Use uint64_t consistently for file sizes in meth1.

+++ implementations/meth1/lib/file.hh
@@ -20,7 +20,7 @@ public:
   void fsync( void );
 
   /* file size */
-  size_t size( void ) const;
+  uint64_t size( void ) const;
 };
 
 #endif /* FILE_HH */
